Jacobians

Basics

The Jacobian of a vector-valued function $f$ is the matrix of all its first-order partial derivatives. Hence, at a point $p = (x_1,...,x_n)$ for $f\colon \mathbb {R} ^{n}\to \mathbb {R}^{m}$ the Jacobian $\mathbf{J}\colon \mathbb {R} ^{n}\to \mathbb {R}^{m \times n}$ is defined as :

\[\begin{aligned} \mathbf{J}(p) = \begin{bmatrix}{\dfrac {\partial f_{1}}{\partial x_{1}}(p)}&\cdots &{\dfrac {\partial f_{1}}{\partial x_{n}}(p)}\\\vdots &\ddots &\vdots \\{\dfrac {\partial f_{m}}{\partial x_{1}}(p)}&\cdots &{\dfrac {\partial f_{m}}{\partial x_{n}}(p)}\end{bmatrix} \end{aligned}\]

JuMPDiff provides different functions to compute the Jacobian of a JuMP model.

Overview

Different dispatches of the function model_jac are available to compute the Jacobian of a specified JuMP model. The basic functionalities are:

JuMPDiff.model_jac — Function

model_jac(model)

This function takes an input a JuMP model and returns the Jacobian matrix of the model constraints w.r.t. all variables within the model.

model_jac(model,vars)

This function takes as input a JuMP model and a set of associated variables and returns the Jacobian matrix of the model constraints w.r.t. to the passed variables.

model_jac(model,vals)

This function takes as input a JuMP model and an array of associated values and returns the Jacobian of the model, evaluated at the specified point. The number of specified values must match exactly the number of variables within the model.

model_jac(model,vars,vals)

This function takes as input a JuMP model and an array of associated variables and returns the Jacobian matrix of the model constraints w.r.t. the passed variables, evaluated at the point specified by vals.

model_jac(model,vals)

This function takes as input a JuMP model and an array of associated values and returns the Jacobian of the model, evaluated at the specified point. The number of specified values must match exactly the number of variables within the model.

model_jac(model,vars,vals)

This function takes as input a JuMP model and a set of associated variables and returns the Jacobian matrix of the model constraints w.r.t. to the passed variables, evaluated at the point specified by vals.

Results retrieved by these methods can also be registered in the model by calling the function model_jac!`.

JuMPDiff.model_jac! — Function

model_jac!(model)

This function takes as input a JuMP model and computes its Jacobian. The Jacobian is registered as expression within the model with the default name "Jacobian".

model_jac!(model,vars)

This function takes as input a JuMP model and an array of associated variables and computes the Jacobian w.r.t. the specified variables. The Jacobian is registered as expression within the model with the default name "Jacobian".

model_jac!(model)

This function takes as input a JuMP model and computes its Jacobian. The Jaocbian is registered as expression within the model with the default name "Jacobian".

model_jac!(model,vars)

This function takes as input a JuMP model and an array of associated variables and computes the Jacobian w.r.t. the specified variables. The Jacobian is registered as expression within the model with the default name "Jacobian".

Examples

The following shall illustrate the use of these methods. Note that the calls of model_jac below are independant of each and shall merely illustrate the different available methods.

julia> model = Model();

julia> @variable(model,x); 

julia> @variable(model,y);

julia> @variable(model,z); 

julia> @constraint(model, 5*x^2 + 3*y^2 == 10);

julia> @constraint(model, x*z == 0); 

julia> @constraint(model, y >= 0);

julia> @objective(model, Max, x^2 - y*x);

julia> model_jac(model)
4×3 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x   0
 0        1    0
 10 x     6 y  0
 z        0    x

julia> model_jac(model,[x,y])
4×2 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x
 0        1
 10 x     6 y
 z        0

julia> model_jac(model,[1,1,1])
4×3 Array{GenericAffExpr{Float64,VariableRef},2}:
 1   -1  0
 0   1   0
 10  6   0
 1   0   1
 
julia> model_jac(model,[y,x], [1,1])
4×2 Array{GenericAffExpr{Float64,VariableRef},2}:
 -1  1
 1   0
 6   10
 0   z

Note that the order in which the partial derivatives are evaluated is the order in which the corresponding variables are loaded into the model, i.e. by the default the first column of the Jacobian will correspond to the variable $x$, the second to $y$ and the third to $z$. The user may revert this order by specifying an array of variables as done in the last function call above. This way it is also possible to only compute the Jacobian w.r.t. certain variables ($y$ and $x$ in the above example). Also note that the number of specified values must always equal the number of specified variables (which is the total number of variables in case no variable array is specified, as done in the third function call). The function will otherwise throw an error:

julia> model_jac(model,[1,1])
ERROR: There must be as many variables as values
[...]

julia> model_jac(model,[x,y,z],[2,2])
ERROR: There must be as many variables as values
[...]

Another thing to note is the order in which the functions in model are evaluated in the Jacobian. From the above example it becomes clear that this order does not equal the order in which the equations were loaded in the model. Rather, the equations are sorted: The first evaluated equation is always the objective function, followed by the constraints. The constraints are evaluated according to the type they represent in the MOI data structure [1]. Indicatively, this means that scalar constraints are evaluated first, and scalar EqualTo constraints are evaluted before GreaterThan constraints. This is followed by the evaluation of the quadratic constraints, where again the EqualTo constraints precede the GreaterThan constraints. Within a single constraint type the evaluation order is the order in which the constraints were loaded to the model. The different types of constraint registered in a JuMP model can always be queried via MOI.get[2]

julia> MOI.get(model,MOI.ListOfConstraints())
2-element Array{Tuple{DataType,DataType},1}:
 (MathOptInterface.ScalarAffineFunction{Float64}, MathOptInterface.GreaterThan{Float64})
 (MathOptInterface.ScalarQuadraticFunction{Float64}, MathOptInterface.EqualTo{Float64})

The model_jac! function extends the previously described functionalities by registering the retrieved results within the JuMP model itself.

julia> model_jac!(model)
4×3 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x   0
 0        1    0
 10 x     6 y  0
 z        0    x

julia> model[:Jacobian]
4×3 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x   0
 0        1    0
 10 x     6 y  0
 z        0    x

julia> model_jac!(model,[x,y],:Jacobian_x_y)
4×2 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x
 0        1
 10 x     6 y
 z        0

julia> model[:Jacobian_x_y]
4×2 Array{GenericAffExpr{Float64,VariableRef},2}:
 2 x - y  -x
 0        1
 10 x     6 y
 z        0

The default name for the registered Jacobian is Jacobian. The result can then be called by querying model[:Jacobian] (where model is the name of the JuMP model). By specifying a different name in the function arguments we can register the result under a different name, and make it callable under that name. This is what we did in the above example, since the second call of model_jac! would have otherwise attempted to re-register the default name Jacobian, resulting in the following error:

julia> model_jac!(model,[x,y])
ERROR: You are trying to register the name Jacobian in the model, which is already occupied. If you have not specified the name Jacobian, then the function you have called assumes Jacobian to be the default name. To overcome this issue specify a different name in the function arguments or unregister the name Jacobian from the model.
[...]

A slightly different error would have been thrown if it were attempted to register the user-defined name twice:

julia> model_jac!(model,[x,y], :Jacobian_x_y)
ERROR: You are trying to register a name which is already occupied in the model. Choose a different one or unregister the existing.
[...]

1Legat, Benoit and Dowson, Oscar and Garcia, Joaquim and Lubin, Miles MathOptInterface: a data structure for mathematical optimization problems. INFORMS Journal on Computing, 2020

2Note that bilinear terms are of quadratic type in JuMP